functions/predict_race_check/requirements.txt
beautifulsoup4==4.8.2
requests==2.23.0
google-cloud-storage==1.26.0
